# Return and track affected and culprit on conflicts